+++ archive_handling_stub.go
@@ -5,7 +5,7 @@ package client
 type AmaasClientArchiveHandler struct {
 }
 
-func (ah *AmaasClientArchiveHandler) initHandler(ac *AmaasClient) error {
+func (ah *AmaasClientArchiveHandler) initHandler(_ *AmaasClient) error {
 	return nil
 }
 
@@ -13,6 +13,6 @@ func (ah *AmaasClientArchiveHandler) archHandlingEnabled() bool {
 	return false
 }
 
-func (ah *AmaasClientArchiveHandler) fileScanRun(fileName string) (string, error) {
+func (ah *AmaasClientArchiveHandler) fileScanRun(_ string) (string, error) {
 	return "", nil
 }